+++ app/models/imports/geo_names.rb
@@ -1,5 +1,0 @@
-module Imports
-  module GeoNames
-    require Rails.root.join("app/lib/imports/geo_names/region_record_connector")
-  end
-end